[ALPS02094832] Thermal driver: Add mt6580
[Detail] Add mt6580 thermal drivers

[Solution] Add mt6580 thermal drivers

[Feature] Thermal driver

MTK-Commit-Id: 410dc0590b4b7c74b34ede4c994642671292d634

CR-Id: ALPS02094832

+++ drivers/misc/mediatek/connectivity/common_detect/mtk_wcn_stub_alps.c
@@ -297,7 +297,7 @@ void mtk_wcn_cmb_stub_func_ctrl(unsigned int type, unsigned int on)
 }
 EXPORT_SYMBOL(mtk_wcn_cmb_stub_func_ctrl);
 
-signed long mtk_wcn_cmb_stub_query_ctrl(void)
+int mtk_wcn_cmb_stub_query_ctrl(void)
 {
 	signed long temp = 0;
